lsp: Move test cases, remove warning filters
